house strnstr

+++ ft_substr.c
@@ -19,7 +19,7 @@ char *ft_substr(const char *s, unsigned int start, size_t len)
     char *new_s;
 
     i = 0;
-    if (start < 0 || start > ft_strlen(s))
+    if ((int)start < 0 || start > ft_strlen(s))
         return NULL;
     new_s = (char *) malloc((len + 1) * sizeof(char));
     while (s[start] && i < len)
